move port to phi/common/
