+++ paddle/fluid/framework/io/shell.h
@@ -38,7 +38,7 @@
 #include <utility>
 #include <vector>
 
-#include "paddle/phi/backends/dynload/port.h"
+#include "paddle/phi/common/port.h"
 #include "paddle/utils/string/string_helper.h"
 
 #if defined(__arm__) || defined(__aarch64__) || defined(__ARM_NEON) || \
