+++ paddle/fluid/inference/analysis/helper.h
@@ -29,7 +29,7 @@ limitations under the License. */
 #include "paddle/fluid/framework/scope.h"
 #include "paddle/fluid/framework/variable.h"
 #include "paddle/fluid/platform/enforce.h"
-#include "paddle/phi/backends/dynload/port.h"
+#include "paddle/phi/common/port.h"
 
 #ifdef _WIN32
 #include <direct.h>
