+++ paddle/fluid/inference/api/helper.h
@@ -35,7 +35,7 @@
 #include "paddle/fluid/memory/stats.h"
 #include "paddle/fluid/platform/enforce.h"
 #include "paddle/fluid/platform/place.h"
-#include "paddle/phi/backends/dynload/port.h"
+#include "paddle/phi/common/port.h"
 #include "paddle/utils/string/printf.h"
 
 extern std::string paddle::framework::DataTypeToString(
