+++ paddle/fluid/platform/timer.h
@@ -15,7 +15,7 @@ limitations under the License. */
 #pragma once
 #include <stdlib.h>
 
-#include "paddle/phi/backends/dynload/port.h"
+#include "paddle/phi/common/port.h"
 #include "paddle/utils/test_macros.h"
 
 #ifdef _WIN32
