+++ paddle/phi/backends/dynload/CMakeLists.txt
@@ -1,5 +1,4 @@
-set(DYNLOAD_COMMON_SRCS dynamic_loader.cc port.cc warpctc.cc warprnnt.cc
-                        lapack.cc)
+set(DYNLOAD_COMMON_SRCS dynamic_loader.cc warpctc.cc warprnnt.cc lapack.cc)
 if(WITH_ASCEND_CL)
   list(REMOVE_ITEM DYNLOAD_COMMON_SRCS warprnnt.cc)
 endif()
