+++ paddle/phi/backends/dynload/cublasLt.h
@@ -22,7 +22,7 @@ limitations under the License. */
 #include <type_traits>
 
 #include "paddle/phi/backends/dynload/dynamic_loader.h"
-#include "paddle/phi/backends/dynload/port.h"
+#include "paddle/phi/common/port.h"
 
 namespace phi {
 namespace dynload {
