+++ paddle/phi/backends/dynload/dynamic_loader.cc
@@ -17,7 +17,7 @@ limitations under the License. */
 #include <string>
 #include <vector>
 #include "paddle/phi/backends/dynload/cupti_lib_path.h"
-#include "paddle/phi/backends/dynload/port.h"
+#include "paddle/phi/common/port.h"
 #include "paddle/phi/core/enforce.h"
 
 #if defined(_WIN32)
