+++ paddle/phi/backends/dynload/flashattn.h
@@ -18,7 +18,7 @@ limitations under the License. */
 
 #include "flashattn/include/flash_attn.h"
 #include "paddle/phi/backends/dynload/dynamic_loader.h"
-#include "paddle/phi/backends/dynload/port.h"
+#include "paddle/phi/common/port.h"
 
 namespace phi {
 namespace dynload {
