+++ paddle/phi/backends/dynload/warprnnt.h
@@ -17,7 +17,7 @@ limitations under the License. */
 #include <mutex>  // NOLINT
 
 #include "paddle/phi/backends/dynload/dynamic_loader.h"
-#include "paddle/phi/backends/dynload/port.h"
+#include "paddle/phi/common/port.h"
 #include "warprnnt/include/rnnt.h"
 
 namespace phi {
