+++ paddle/phi/common/CMakeLists.txt
@@ -1,1 +1,8 @@
-collect_srcs(common_srcs SRCS place.cc scalar.cc int_array.cc memory_utils.cc)
+collect_srcs(
+  common_srcs
+  SRCS
+  place.cc
+  scalar.cc
+  int_array.cc
+  memory_utils.cc
+  port.cc)
