+++ test/cpp/inference/test_helper.h
@@ -24,7 +24,7 @@ limitations under the License. */
 #include "paddle/fluid/inference/io.h"
 #include "paddle/fluid/platform/errors.h"
 #include "paddle/fluid/platform/profiler.h"
-#include "paddle/phi/backends/dynload/port.h"
+#include "paddle/phi/common/port.h"
 
 COMMON_DECLARE_bool(use_mkldnn);
