+++ test/cpp/phi/kernels/test_cpu_vec.cc
@@ -18,7 +18,7 @@ limitations under the License. */
 
 #include "glog/logging.h"
 #include "gtest/gtest.h"
-#include "paddle/phi/backends/dynload/port.h"
+#include "paddle/phi/common/port.h"
 #include "paddle/phi/kernels/funcs/cpu_vec.h"
 
 namespace phi {
